change buttons design on charger card and charger details, add error state, error handle delete charger
