+++ src/components/Charger.js
@@ -5,6 +5,7 @@ import {
   CardMedia,
   CardActions,
   CardActionArea,
+  Button
 } from "@mui/material";
 import { useNavigate } from "react-router-dom";
 import { ButtonGroup } from "./ButtonGroup";
@@ -43,7 +44,16 @@ export const Charger = ({ charger }) => {
         </CardContent>
       </CardActionArea>
       <CardActions>
-        <ButtonGroup key={charger.id} charger={charger} />
+        <Button
+          type="submit"
+          value="active"
+          variant="contained"
+          onClick={handleClick}
+          style={{ marginRight: "16px" }}
+        >
+          View
+        </Button>
+        {/* <ButtonGroup key={charger.id} charger={charger} /> */}
       </CardActions>
     </Card>
   );
